backend/requirements.txt
# Web framework & server
fastapi>=0.115.0
uvicorn[standard]>=0.32.0

# Database
sqlalchemy>=2.0.36

# Validation & serialization
pydantic[email]>=2.10.0
email-validator>=2.0.0

# Form & file upload
python-multipart>=0.0.12

# Auth
pyotp>=2.9.0
passlib[bcrypt]>=1.7.4
python-jose[cryptography]>=3.5.0
bcrypt>=4.2.0

# HTTP client
requests>=2.32.0
